backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlalchemy==2.0.36
beautifulsoup4==4.12.3
requests==2.32.3
boto3==1.35.0
pydantic==2.9.0
python-multipart==0.0.12
pymupdf==1.25.3
faiss-cpu==1.9.0.post1
langchain==0.3.14
langchain-text-splitters==0.3.4
tiktoken==0.8.0
schedule==1.2.0
openai==1.54.0
httpx==0.27.0
pyyaml==6.0.1
